+++ app/src/main/java/com/example/rankingu/ui/SignUpActivities/SignUp_Estudiante.java
@@ -1,4 +1,4 @@
-package com.example.rankingu.ui;
+package com.example.rankingu.ui.SignUpActivities;
 
 import android.os.Bundle;
