+++ app/src/main/res/layout/activity_sign_up.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.SignUp">
+    tools:context=".ui.SignUpActivities.SignUp">
 
     <TextView
         android:id="@+id/textView"
