+++ app/src/main/res/layout/activity_sign_up__estudiante.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.SignUp_Estudiante">
+    tools:context=".ui.SignUpActivities.SignUp_Estudiante">
 
     <TextView
         android:id="@+id/textView16"
